Refactor: 게시글 보드 로직 훅으로 분리, 페이지셀렉터 합침, 스타일 분리
