+++ src/components/pages/Mypage/hooks/index.js
@@ -1,4 +1,5 @@
 import useMypageProfile from './useMypageProfile';
 import useProfileSection from './useProfileSection';
+import useMyArticleBoard from './useMyArticleBoard';
 
-export { useMypageProfile, useProfileSection };
+export { useMypageProfile, useProfileSection, useMyArticleBoard };
